refactor: crate::utils::json::GoVecVisitor -> crate::json::vec::GoVecVisitor
