+++ src/beacon/beacon_entries.rs
@@ -83,7 +83,7 @@ pub mod json {
     }
 
     pub mod vec {
-        use crate::utils::json::GoVecVisitor;
+        use crate::json::vec::GoVecVisitor;
         use serde::ser::SerializeSeq;
 
         use super::*;
