+++ src/blocks/header/json.rs
@@ -152,7 +152,7 @@ where
 }
 
 pub mod vec {
-    use crate::utils::json::GoVecVisitor;
+    use crate::json::vec::GoVecVisitor;
     use serde::ser::SerializeSeq;
 
     use super::*;
