+++ src/json/cid.rs
@@ -43,7 +43,7 @@ struct CidMap {
 }
 
 pub mod vec {
-    use crate::utils::json::GoVecVisitor;
+    use crate::json::vec::GoVecVisitor;
     use serde::ser::SerializeSeq;
 
     use super::*;
