+++ src/json/mod.rs
@@ -10,7 +10,9 @@ pub mod sector;
 pub mod signature;
 pub mod signed_message;
 pub mod token_amount;
+pub mod vec;
 pub mod vrf;
+
 #[cfg(test)]
 mod tests {
     mod address_test;
