+++ src/json/signed_message.rs
@@ -69,7 +69,7 @@ pub mod json {
     }
 
     pub mod vec {
-        use crate::utils::json::GoVecVisitor;
+        use crate::json::vec::GoVecVisitor;
         use serde::ser::SerializeSeq;
 
         use super::*;
